Add Redis, bull queue and job
