+++ server/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "bull": "^3.13.0",
     "cookie-parser": "~1.4.4",
     "csv-parser": "^2.3.2",
     "debug": "~2.6.9",
